logs: Add separate install/uninstall logs
To avoid confusion as to what the logs command does, we separate into
distinct 'install' and 'uninstall' logs.
